player-verification-system/requirements.txt
Flask==2.3.0
flask-socketio==5.3.0
opencv-python==4.8.0.74
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
werkzeug==2.3.0
python-socketio==5.9.0
eventlet==0.33.3
